+++ packages/responses/src/Domain/Temp/Subscription.ts
@@ -1,0 +1,7 @@
+export type Subscription = {
+  planName: string
+  endsAt: number
+  createdAt: number
+  updatedAt: number
+  cancelled: boolean
+}