+++ packages/responses/src/Domain/Temp/ValetTokenOperation.ts
@@ -1,0 +1,6 @@
+export enum ValetTokenOperation {
+  Read = 'read',
+  Write = 'write',
+  Delete = 'delete',
+  Move = 'move',
+}