+++ packages/clipper/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.98](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.97...@standardnotes/clipper@1.1.98) (2023-07-17)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.97](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.96...@standardnotes/clipper@1.1.97) (2023-07-16)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.96](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.95...@standardnotes/clipper@1.1.96) (2023-07-16)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.95](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.94...@standardnotes/clipper@1.1.95) (2023-07-15)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.94](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.93...@standardnotes/clipper@1.1.94) (2023-07-15)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
+## [1.1.93](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.92...@standardnotes/clipper@1.1.93) (2023-07-15)
+
+**Note:** Version bump only for package @standardnotes/clipper
+
 ## [1.1.92](https://github.com/standardnotes/app/compare/@standardnotes/clipper@1.1.91...@standardnotes/clipper@1.1.92) (2023-07-14)
 
 **Note:** Version bump only for package @standardnotes/clipper