+++ packages/encryption/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.21.52](https://github.com/standardnotes/app/compare/@standardnotes/encryption@1.21.51...@standardnotes/encryption@1.21.52) (2023-07-17)
+
+**Note:** Version bump only for package @standardnotes/encryption
+
 ## [1.21.51](https://github.com/standardnotes/app/compare/@standardnotes/encryption@1.21.50...@standardnotes/encryption@1.21.51) (2023-07-13)
 
 **Note:** Version bump only for package @standardnotes/encryption