+++ packages/filepicker/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.28.61](https://github.com/standardnotes/app/compare/@standardnotes/filepicker@1.28.60...@standardnotes/filepicker@1.28.61) (2023-07-17)
+
+**Note:** Version bump only for package @standardnotes/filepicker
+
 ## [1.28.60](https://github.com/standardnotes/app/compare/@standardnotes/filepicker@1.28.59...@standardnotes/filepicker@1.28.60) (2023-07-13)
 
 **Note:** Version bump only for package @standardnotes/filepicker