+++ packages/responses/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.30](https://github.com/standardnotes/app/compare/@standardnotes/responses@1.13.29...@standardnotes/responses@1.13.30) (2023-07-17)
+
+**Note:** Version bump only for package @standardnotes/responses
+
 ## [1.13.29](https://github.com/standardnotes/app/compare/@standardnotes/responses@1.13.28...@standardnotes/responses@1.13.29) (2023-07-13)
 
 **Note:** Version bump only for package @standardnotes/responses