+++ packages/web/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.166.16](https://github.com/standardnotes/app/compare/@standardnotes/web@3.166.15...@standardnotes/web@3.166.16) (2023-07-17)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.166.15](https://github.com/standardnotes/app/compare/@standardnotes/web@3.166.14...@standardnotes/web@3.166.15) (2023-07-16)
+
+### Bug Fixes
+
+* Fixed issue where some editors would not open on Android ([5932eb8](https://github.com/standardnotes/app/commit/5932eb814cbf81568d435afe134b99441d71f930))
+
+## [3.166.14](https://github.com/standardnotes/app/compare/@standardnotes/web@3.166.13...@standardnotes/web@3.166.14) (2023-07-16)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.166.13](https://github.com/standardnotes/app/compare/@standardnotes/web@3.166.12...@standardnotes/web@3.166.13) (2023-07-15)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.166.12](https://github.com/standardnotes/app/compare/@standardnotes/web@3.166.11...@standardnotes/web@3.166.12) (2023-07-15)
+
+**Note:** Version bump only for package @standardnotes/web
+
+## [3.166.11](https://github.com/standardnotes/app/compare/@standardnotes/web@3.166.10...@standardnotes/web@3.166.11) (2023-07-15)
+
+**Note:** Version bump only for package @standardnotes/web
+
 ## [3.166.10](https://github.com/standardnotes/app/compare/@standardnotes/web@3.166.9...@standardnotes/web@3.166.10) (2023-07-14)
 
 **Note:** Version bump only for package @standardnotes/web
@@ -116,8 +142,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 **Note:** Version bump only for package @standardnotes/web
 
 # [3.164.0](https://github.com/standardnotes/app/compare/@standardnotes/web@3.163.3...@standardnotes/web@3.164.0) (2023-07-03)
-
-### Features
 
 **Note:** Version bump only for package @standardnotes/web